% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baselayer}
\alias{baselayer}
\title{Baselayer data}
\format{
A \verb{Large sfc_Multipolygon}
}
\usage{
baselayer
}
\description{
Baselayer data of Denmark. From \url{https://dawadocs.dataforsyningen.dk/dok/dagi}
}
\keyword{datasets}
