% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_getAdmin.R
\name{cc_getAdmin}
\alias{cc_getAdmin}
\title{Get lorem ipsum}
\usage{
cc_getAdmin(spatial = TRUE, level = NULL)
}
\arguments{
\item{spatial}{a logical value. Return data of class \code{sf}? Defaults to \code{TRUE}}

\item{level}{lorem ipsum}
}
\value{
data lorem ipsum
}
\description{
lorem ipsum
}
\author{
Jeppe Vierø
}
