% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_readBaseLayer.R
\name{cc_readBaseLayer}
\alias{cc_readBaseLayer}
\title{Get lorem ipsum}
\usage{
cc_readBaseLayer(as_sf = TRUE)
}
\arguments{
\item{as_sf}{A logical value. Return data as an object of class \code{sf}? Default is TRUE. If FALSE, imports as a GEOJSON of class \code{geomultipolygon}}
}
\value{
data lorem ipsum
}
\description{
lorem ipsum
}
\author{
Jeppe Vierø
}
