% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_getAdmin.R
\name{cc_getAdmin}
\alias{cc_getAdmin}
\title{Get voting station data}
\usage{
cc_getAdmin(spatial = TRUE, level = NULL)
}
\arguments{
\item{spatial}{a logical value. Return data of class \code{sf}? Defaults to \code{TRUE}}
}
\value{
data on danish voting stations
}
\description{
This function extracts spatial data on voting stations in Denmark from \url{https://info.aws.dk/afstemningsomraader/}
}
\examples{
df <- cc_getAdmin(level == "region")
}
\author{
Jeppe Vierø
}
